+++ changelog.md
@@ -26,4 +26,4 @@
 
 - [GLB, RNM] remove `core` module and move its content (submodules) into the root of the package
 
-- [GLB, TST] add folder `test` (draft version) with unittests. For example, basic tests for the `maxvol` module have been added, in the future all modules and functions will be covered by tests
+- [GLB, TST] add folder `test` (draft version) with unittests. For example, basic tests for the `maxvol` module have been added, later all modules and functions will be covered by tests